Documentation

Public API

Documentation for ProfileLikelihood.jl's public functions.

Error/objective functions

ProfileLikelihood.relative_error — Method

relative_error(data::AbstractVector{<:Real}, sol::AbstractVector{<:Real}, 
               noise_level::Real)

The objective function for when the noise in the data follow relative error.

The objective function is

$\hat{θ} = \arg \min_{θ} \frac{1}{η^2} \sum_{i=1}^{n} \left( \frac{y_i - g(t_i,θ)}{g(t_i,θ)} \right)^2 + 2 \sum_{i=1}^n \log g(t_i,θ)$

where η is the noise level, $g(t_i,θ)$ is the $i$th point of the predicted solution, and $y_i$ is the $i$th data point.

ProfileLikelihood.poisson_error — Method

poisson_error(data::AbstractVector{<:Real}, sol::AbstractVector{<:Real})

The objective function for when the data follow a Poisson distribution.

The objective function is

$\hat{θ} = \arg \min_{θ} -\sum_{i=1}^n \log (g(t_i,u,θ)^{y_i}) + \sum_{i=1}^n g(t_i,u,θ)$

where $g(t_i,θ)$ is the $i$th point of the predicted solution and $y_i$ is the $i$th data point.

ProfileLikelihood.const_variance_error — Method

const_variance_error(data::AbstractVector{<:Real}, sol::AbstractVector{<:Real}, 
                     sigma::Real)

The objective function for when the data follow a normal distribution with mean 0 and known variance $\sigma^2$.

The objective function is

$\hat{θ} = \arg \min_{θ} \sum_{i=1}^n (y_i - g(t_i,u,θ))^2 / \sigma^2$

where $g(t_i,θ)$ is the $i$th point of the predicted solution and $y_i$ is the $i$th data point.

ProfileLikelihood.likelihood_const — Method

likelihood_const(obj::String; 
                 noise_level::Real=0.0, 
                 times::AbstractVector{<:Real}=Vector{Real}(), 
                 data::AbstractVector{<:Real}=Vector{Real}(), 
                 sigma::Real = -1.0)

This gives the constant in the likelihood function when the error is assumed to follow relative error or constant variance error or when the data is assumed to follow a Poisson distribution.

The constant for the likelihood function is often dropped when finding the minimizer of the likelihood function. We typically add this back when generating the profile likelihood plots. For relative error, the constant is

$n \log η^2 + n \log 2 π$

where $η$ is the noise level and $n$ is the number of data points. For constant variance error, the constant is

$n \log 2π + n \log σ^2$

where $σ$ is the standard deviation of the error and $n$ is the number of data points. For poisson error, the constant is

$\sum_{i=1}^n \log (y_i!)$

where $n$ is the number of data points and $y_i$ is the $i$th data point of the observable $y$.

Generate data

ProfileLikelihood.generate_data — Method

generate_data(index::Integer, seed::Integer, dist::Function, 
              prob::SciMLBase.AbstractDEProblem, 
              alg::SciMLBase.AbstractDEAlgorithm, 
              times::AbstractVector{<:Real}; 
              incidence_obs_status::Bool = false, 
              kwargs...)

This generates perfect and noisy data and incidence data of the state variables of the system of differential equations.

Let $\mathcal{I}:[t_0,t_1,\dots,t_n] \rightarrow \mathbb{R}$ be the number of incidences defined by $0$ if $t=t_0$ and $\mathcal{C}(t_i) - \mathcal{C}(t_{i-1})$ if $t \neq t_0$ where $\mathcal{C}$ is cumulative data. Note that times is $[t_0,t_1,\dots,t_n]$.

Arguments

index::Integer: Index of the component of the state variables for data collection.

seed::Integer: Seed of RNG generator. Set for reproducible results.

dist::Function: Function to generate noisy data. The inputs are the data points sampled at points in times.

prob::SciMLBase.AbstractDEProblem: DE Problem (see DifferentialEquations.jl for more information).

alg::SciMLBase.AbstractDEAlgorithm: DE Solver Algorithm (see DifferentialEquations.jl for more information).

times::AbstractVector{<:Real}: Times that the data points will be sampled at.

Keywords

incidence_obs_status::Bool = false: Determine whether the data is incidence data or not. If true, the indexth state variable must be cumulative data.

kwargs...: Keyword arguments to be passed into the DE solver. See DifferentialEquations.jl's Common Solver Options.

Returns

perfect_data: Data without noise.

noisy_data: Data with noise according to dist.

ProfileLikelihood.generate_incidence_data — Method

generate_incidence_data(index::Integer, 
                        prob::SciMLBase.AbstractDEProblem, 
                        alg::SciMLBase.AbstractDEAlgorithm, 
                        times::AbstractVector{<:Real}; 
                        kwargs...)

This generates incidence data without noise by solving the DEs and computing incidence data from cumulative data according to the indexth state variable.

Let $\mathcal{I}:[t_0,t_1,\dots,t_n] \rightarrow \mathbb{R}$ be the number of incidences defined by $0$ if $t=t_0$ and $\mathcal{C}(t_i) - \mathcal{C}(t_{i-1})$ if $t \neq t_0$ where $\mathcal{C}$ is cumulative data. Note that times is $[t_0,t_1,\dots,t_n]$.

Arguments

index::Integer: Index of the component of the state variables for data collection.

prob::SciMLBase.AbstractDEProblem: DE Problem (see DifferentialEquations.jl for more information).

alg::SciMLBase.AbstractDEAlgorithm: DE Solver Algorithm (see DifferentialEquations.jl for more information).

times::AbstractVector{<:Real}: Times that the data points will be sampled at.

Keywords

kwargs...: Keyword arguments to be passed into the DE solver. See DifferentialEquations.jl's Common Solver Options.

Returns

incidence_data: Incidence data of the indexth state variable which represents cumulative data.

Estimating parameters

ProfileLikelihood.estimate_params — Function

estimate_params(p0::AbstractVector{<:Real}, 
                data::AbstractVector{<:AbstractVector{<:Real}}, 
                sol_obs::AbstractVector{<:Integer}, 
                prob::SciMLBase.AbstractDEProblem, 
                alg_diff::SciMLBase.AbstractDEAlgorithm, 
                times::AbstractVector{<:Real}, obj_arr::AbstractVector{<:Function}, 
                alg_opti, 
                lb::AbstractVector{<:Real}, ub::AbstractVector{<:Real}; 
                incidence_obs::AbstractVector{<:Int}=[], param_index::Int=0, 
                param_eval::Real=0.0, 
                solver_diff_opts::Dict=Dict(), opti_prob_opts::Dict=Dict(), 
                opti_solver_opts::Dict=Dict(), print_status::Bool=false)

This estimate the parameters of the system of differential equations using the objective functions in obj_arr given prob using the optimization algorithm alg_opti.

Estimation of parameters is done using the objective function array obj_arr where each objective function in the array corresponds to data described by sol_obs and incidence_obs. For instance, if obj_arr = [obj1 obj2],sol_obs = [2], incidence_obs = [3], then obj1 use data corresponding to the second state variable of the DEs and obj2 use incidence data of the third state variable of the DEs. Noisy data or any other type of data should be passed to data and in the same order as in the arrays sol_obs and incidence_obs. Fixing a parameter can be done using param_index. Instead, this return the fitted parameters with param_eval for the fixed parameter and accordingly, the loss is the objective function evaluated at the fitted parameters with param_eval for the fixed parameter.

Arguments

p0::AbstractVector{<:Real}: Starting guess for optimization. If param_index and param_eval are used, then p0 should not contain the fixed parameter.

data::AbstractVector{<:AbstractVector{<:Real}}: Vector of data used for optimizing parameters. data must be in the same order as in sol_obs and incidence_obs.

sol_obs::AbstractVector{<:Integer}: Indices of the state variables of the DEs to be used for sampling data points.

prob::SciMLBase.AbstractDEProblem: DE Problem (see DifferentialEquations.jl for more information).

alg_diff::SciMLBase.AbstractDEAlgorithm: DE Solver Algorithm (see DifferentialEquations.jl for more information).

times::AbstractVector{<:Real}: Times that the data points will be sampled at.

obj_arr::AbstractVector{Function}: Vector of objective functions.

alg_opti: Optimization algorithm (see Optimization.jl for a list of algorithms that could be used).

lb::AbstractVector{<:Real}: Lower bound (does not need to be changed if param_index and param_eval are used).

ub::AbstractVector{<:Real}: Upper bound (does not need to be changed if param_index and param_eval are used).

Keywords

incidence_obs::AbstractVector{<:Int}=[]: Indices of the state variables of the DEs to find incidence data of. The state variables must be cumulative data.

param_index::Int=0: Index of the parameter vector that is fixed.

param_eval::Real=0.0: Value for the fixed parameter.

solver_diff_opts::Dict=Dict(): Keyword arguments to be passed into the DE solver. See DifferentialEquations.jl's Common Solver Options.

opti_prob_opts::Dict=Dict(): Keyword arguments to be passed into the optimization problem. See Optimization.jl's Defining OptimizationProblems.

opti_solver_opts::Dict=Dict(): Keyword arguments to be passed into the optimization solver. See Optimization.jl's Common Solver Options.

print_status::Bool=false: Determine whether the orginal output of the optimization algorithm is printed or not.

Returns

minimum: Loss of the fitted parameters according to the objective function(s).

minimizer: Fitted parameters.

ProfileLikelihood.estimate_params_multistart — Function

estimate_params_multistart(p0::AbstractVector{<:Real}, 
                           data::AbstractVector{<:AbstractVector{<:Real}}, 
                           sol_obs::AbstractVector{<:Integer}, 
                           prob::SciMLBase.AbstractDEProblem, 
                           alg_diff::SciMLBase.AbstractDEAlgorithm, 
                           times::AbstractVector{<:Real}, 
                           obj_arr::AbstractVector{<:Function}, 
                           alg_opti, 
                           alg_opti_local,
                           lb::AbstractVector{<:Real}, ub::AbstractVector{<:Real}; 
                           incidence_obs::AbstractVector{<:Int}=[], param_index::Int=0, 
                           param_eval::Real=0.0, 
                           solver_diff_opts::Dict=Dict(), opti_prob_opts::Dict=Dict(), 
                           opti_solver_opts::Dict=Dict(), print_status::Bool=false)

This estimate the parameters of the system of differential equations using the objective function obj_arr given prob using the multi-start optimization algorithm alg_opti from MultiStartOptimization and the local optimization algorithm alg_opti_local from the NLopt.jl.

The algorithm alg_opti is always MultistartOptimization.TikTak(n) where n is the number of starting points generated from the Sobol sequence. MultistartOptimization.jl only support optimziation algorithms from NLopt.jl. The recommended local optimization algorithm is NLopt.LN_NELDERMEAD(). See the docstring for estimate_params for more information for what estimate_params_multistart do.

Arguments

p0::AbstractVector{<:Real}: Starting guess for optimization. If param_index and param_eval are used, then p0 should not contain the fixed parameter.

data::AbstractVector{<:AbstractVector{<:Real}}: Vector of data used for optimizing parameters. data must be in the same order as in sol_obs and incidence_obs.

sol_obs::AbstractVector{<:Integer}: Indices of the state variables of the DEs to be used for sampling data points.

prob::SciMLBase.AbstractDEProblem: DE Problem (see DifferentialEquations.jl for more information).

alg_diff::SciMLBase.AbstractDEAlgorithm: DE Solver Algorithm (see DifferentialEquations.jl for more information).

times::AbstractVector{<:Real}: Times that the data points will be sampled at.

obj_arr::AbstractVector: Vector of objective functions.

alg_opti: Global optimization algorithm. Typically, MultistartOptimization.TikTak(n) where n is the number of starting points generated from the Sobol sequence.

alg_opti_local: Local optimization algorithm. Must be an algorithm from NLopt.jl.

lb::AbstractVector{<:Real}: Lower bound (does not need to be changed if param_index and param_eval are used).

ub::AbstractVector{<:Real}: Upper bound (does not need to be changed if param_index and param_eval are used).

Keywords

incidence_obs::AbstractVector{<:Int}=[]: Indices of the state variables of the DEs to find incidence data of which is used for sampling data points. The state variables must be cumulative data.

param_index::Int=0: Index of the parameter vector that is fixed.

param_eval::Real=0.0: Value for the fixed parameter.

solver_diff_opts::Dict=Dict(): Keyword arguments to be passed into the DE solver. See DifferentialEquations.jl's Common Solver Options.

opti_prob_opts::Dict=Dict(): Keyword arguments to be passed into the optimization problem. See Optimization.jl's Defining OptimizationProblems.

opti_solver_opts::Dict=Dict(): Keyword arguments to be passed into the optimization solver. See Optimization.jl's Common Solver Options.

print_status::Bool=false: Determine whether the orginal output of the optimization algorithm is printed or not.

Returns

minimum: Loss of the fitted parameters according to the objective function(s).

minimizer: Fitted parameters.

Likelihood function

ProfileLikelihood.likelihood — Function

likelihood(params::AbstractVector{<:Real}, 
           data::AbstractVector{<:AbstractVector{<:Real}},
           sol_obs::AbstractVector{<:Integer}, 
           prob::SciMLBase.AbstractDEProblem, 
           alg::SciMLBase.AbstractDEAlgorithm, 
           times::AbstractVector{<:Real}, 
           obj_arr::AbstractVector{<:Function}; 
           incidence_obs::AbstractVector{<:Integer} = Int64[], 
           param_index::Integer=0, 
           param_eval::Real=0.0, 
           solver_diff_opts::Dict = Dict())::Float64

This computes the likelihood based on the given parameters params, data, and likelihood functions in obj_arr.

Each objective function in obj_arr corresponds to data described by sol_obs and incidence_obs. For instance, if obj_arr = [obj1 obj2],sol_obs = [2], incidence_obs = [3], then obj1 use data corresponding to the second state variable of the DEs and obj2 use incidence data of the third state variable of the DEs. Noisy data or any other type of data should be passed to data and in the same order as in the arrays sol_obs and incidence_obs. Fixing a parameter can be done using param_index. Instead, this return the fitted parameters with param_eval for the fixed parameter and accordingly, the loss is the objective function evaluated at the fitted parameters with param_eval for the fixed parameter.

Arguments

params::AbstractVector{<:Real}: Parameters to evaluate at. If param_index and param_eval are used, then params should not contain the fixed parameter.

data::AbstractVector{<:AbstractVector{<:Real}}: Vector of data used for computing likelihood. data must be in the same order as in sol_obs and incidence_obs.

sol_obs::AbstractVector{<:Integer}: Indices of the state variables of the DEs to be used for sampling data points.

prob::SciMLBase.AbstractDEProblem: DE Problem (see DifferentialEquations.jl for more information).

alg::SciMLBase.AbstractDEAlgorithm: DE Solver Algorithm (see DifferentialEquations.jl for more information).

times::AbstractVector{<:Real}: Times that the data points will be sampled at.

obj_arr::AbstractVector: Vector of objective functions.

Keywords

incidence_obs::AbstractVector{<:Integer} = []: Indices of the state variables of the DEs to find incidence data of. The state variables must be cumulative data.

param_index::Integer=0: Index of the parameter vector that is fixed.

param_eval::Real=0.0: Value for the fixed parameter.

solver_diff_opts::Dict = Dict(): Keyword arguments to be passed into the DE solver. See DifferentialEquations.jl's Common Solver Options.

Returns

loss: Likelihood of observing data given params.

Threshold

ProfileLikelihood.find_threshold — Method

find_threshold(confidence::Real, nums_params::Integer, loss::Real)

This finds the threshold used for calculating likelihood-based confidence intervals.

The threshold Δ_α is the α quantile of the χ^2 distribution with df=1 for pointwise confidence intervals and df=# of parameters for simulataneous confidence intervals.

Arguments

confidence::Real: Level of confidence used for confidence interval.

nums_params::Integer: Number of parameters. $df$ = 1 for pointwise confidence intervals and $df$ = number of unknown parameters for simulataneous confidence intervals.

loss:Real: Minimum loss according to negative log-likelihood function after optimization.

Return

threshold: Threshold used for likelihood-based confidence interval.

Finding profile likelihood

ProfileLikelihood.find_profile_likelihood — Function

find_profile_likelihood(step_size::Real, max_steps::Integer, param_index::Integer,
                        param_fitted::AbstractVector{<:Real}, 
                        data::AbstractVector{<:AbstractVector{<:Real}}, 
                        sol_obs::AbstractVector{<:Integer}, 
                        threshold::Real, loss::Real, 
                        prob::SciMLBase.AbstractDEProblem, 
                        alg_diff::SciMLBase.AbstractDEAlgorithm,
                        times::AbstractVector{<:Real},
                        obj_arr::AbstractVector, 
                        alg_opti, 
                        lb::AbstractVector{<:Real}, ub::AbstractVector{<:Real}; 
                        incidence_obs::AbstractVector{<:Integer}=Int64[], 
                        solver_diff_opts::Dict=Dict(), opti_prob_opts::Dict=Dict(), 
                        opti_solver_opts::Dict=Dict(), 
                        print_status::Bool=false, pl_const::Real=0.0)

This find the profile likelihood of the param_indexth parameter in param_fitted.

This implements a simple fixed step size algorithm to compute profile likelihood. It start at the minimum point of the profile likelihood and take step in the right direction and left direction. The algorithm stop taking steps if the maximum allowed steps max_step is reached or if the points being computed is above the threshold.

Arguments

step_size::Real: Step size to take for computing points.

max_steps::Integer: Maximum steps allowed.

param_index::Integer: Index of the parameter vector that is fixed.

param_fitted::AbstractVector{<:Real}: Optimized parameters. Should be the return vector minimizer of estimate_params or estimate_params_multistart.

data::AbstractVector{<:AbstractVector{<:Real}}: Vector of data used for optimizing parameters. data must be in the same order as in sol_obs and incidence_obs.

sol_obs::AbstractVector{<:Integer}: Indices of the state variables of the DEs to be used for sampling data points.

threshold::Real: Value for which the algorithm will stop computing points if the loss of the computed points is greater than the threshold. Does not necessarily need to be the same threshold as the threshold value in find_threshold.

loss::Real: Loss of the fitted parameters according to the objective function(s). Should be the return value minimum of estimate_params or estimate_params_multistart.

prob::SciMLBase.AbstractDEProblem: DE Problem (see DifferentialEquations.jl for more information).

alg_diff::SciMLBase.AbstractDEAlgorithm: DE Solver Algorithm (see DifferentialEquations.jl for more information).

times::AbstractVector{<:Real}: Times that the data points will be sampled at.

obj_arr::AbstractVector{<:Function}: Vector of objective functions.

alg_opti: Optimization algorithm (see Optimization.jl for a list of algorithms that could be used).

lb::AbstractVector{<:Real}: Lower bound (does not need to be changed if param_index and param_eval are used).

ub::AbstractVector{<:Real}: Upper bound (does not need to be changed if param_index and param_eval are used).

Keywords

incidence_obs::AbstractVector{<:Integer}=Int64[]: Indices of the state variables of the DEs to find incidence data of. The state variables must be cumulative data.

solver_diff_opts::Dict=Dict(): Keyword arguments to be passed into the DE solver. See DifferentialEquations.jl's Common Solver Options.

opti_prob_opts::Dict=Dict(): Keyword arguments to be passed into the optimization problem. See Optimization.jl's Defining OptimizationProblems.

opti_solver_opts::Dict=Dict(): Keyword arguments to be passed into the optimization solver. See Optimization.jl's Common Solver Options.

print_status::Bool=false: Determine whether the orginal output of the optimization algorithm is printed or not.

pl_const::Real=0.0: Constant that is added to loss of the computed points. The constant can be computed by likelihood_const.

Returns

theta: Values of the fixed parameter that is explored.

sol: Value of likelihood computed at the fixed parameter where the other parameters are optimized.

ProfileLikelihood.find_profile_likelihood_multistart — Function

find_profile_likelihood_multistart(step_size::Real, max_steps::Integer, param_index::Integer, 
                                   param_fitted::AbstractVector{<:Real}, 
                                   data::AbstractVector{<:AbstractVector{<:Real}}, 
                                   sol_obs::AbstractVector{<:Integer}, 
                                   threshold::Real, loss::Real, 
                                   prob::SciMLBase.AbstractDEProblem, 
                                   alg_diff::SciMLBase.AbstractDEAlgorithm,  
                                   times::AbstractVector{<:Real}, 
                                   obj_arr::AbstractVector{<:Function}, 
                                   alg_opti, alg_opti_local, 
                                   lb::AbstractVector{<:Real}, ub::AbstractVector{<:Real}; 
                                   incidence_obs::AbstractVector{<:Integer}=Int64[], 
                                   solver_diff_opts=Dict(), opti_prob_opts=Dict(), 
                                   opti_solver_opts=Dict(), 
                                   print_status::Bool=false, pl_const::Real=0.0)

This find the profile likelihood of the param_indexth parameter in param_fitted.

This implements a simple fixed step size algorithm to compute profile likelihood. It start at the minimum point of the profile likelihood and take step in the right direction and left direction. The algorithm stop taking steps if the maximum allowed steps max_step is reached or if the points being computed is above the threshold. The difference between find_profile_likelihood and find_profile_likelihood_multistart is that this function uses the multi-start optimization algorithm alg_opti from MultiStartOptimization and the local optimization algorithm alg_opti_local from the NLopt.jl. This is similar to the differences between estimate_params and estimate_params_multistart.

Arguments

step_size::Real: Step size to take for computing points.

max_steps::Integer: Maximum steps allowed.

param_index::Integer: Index of the parameter vector that is fixed.

param_fitted::AbstractVector{<:Real}: Optimized parameters. Should be the return vector minimizer of estimate_params or estimate_params_multistart.

data::AbstractVector{<:AbstractVector{<:Real}}: Vector of data used for optimizing parameters. data must be in the same order as in sol_obs and incidence_obs.

sol_obs::AbstractVector{<:Integer}: Indices of the state variables of the DEs to be used for sampling data points.

threshold::Real: Value for which the algorithm will stop computing points if the loss of the computed points is greater than the threshold. Does not necessarily need to be the same threshold as the threshold value in find_threshold.

loss::Real: Loss of the fitted parameters according to the objective function(s). Should be the return value minimum of estimate_params or estimate_params_multistart.

prob::SciMLBase.AbstractDEProblem: DE Problem (see DifferentialEquations.jl for more information).

alg_diff::SciMLBase.AbstractDEAlgorithm: DE Solver Algorithm (see DifferentialEquations.jl for more information).

times::AbstractVector{<:Real}: Times that the data points will be sampled at.

obj_arr::AbstractVector{<:Function}: Vector of objective functions.

alg_opti: Global optimization algorithm. Typically, MultistartOptimization.TikTak(n) where n is the number of starting points generated from the Sobol sequence.

alg_opti_local: Local optimization algorithm. Must be an algorithm from NLopt.jl.

lb::AbstractVector{<:Real}: Lower bound (does not need to be changed if param_index and param_eval are used).

ub::AbstractVector{<:Real}: Upper bound (does not need to be changed if param_index and param_eval are used).

Keywords

incidence_obs::AbstractVector{<:Integer}=Int64[]: Indices of the state variables of the DEs to find incidence data of. The state variables must be cumulative data.

solver_diff_opts::Dict=Dict(): Keyword arguments to be passed into the DE solver. See DifferentialEquations.jl's Common Solver Options.

opti_prob_opts::Dict=Dict(): Keyword arguments to be passed into the optimization problem. See Optimization.jl's Defining OptimizationProblems.

opti_solver_opts::Dict=Dict(): Keyword arguments to be passed into the optimization solver. See Optimization.jl's Common Solver Options.

print_status::Bool=false: Determine whether the orginal output of the optimization algorithm is printed or not.

pl_const::Real=0.0: Constant that is added to loss of the computed points. The constant can be computed by likelihood_const.

Returns

theta: Values of the fixed parameter that is explored.

sol: Value of likelihood computed at the fixed parameter where the other parameters are optimized.

Processing data

ProfileLikelihood.find_roots — Function

find_roots(interpolation, theta::AbstractVector{<:Real}, threshold::Real)

This finds the intercepts of the threshold and $χ^2$ which corresponds to the confidence intervals.

interpolation is a function which takes in a single input

Arguments

interpolation: Function which takes in a single input $x$ and output $χ^2(x)$ where $χ^2$ is profile likelihood. This should be an interpolation of the points found using PL.

theta::AbstractVector{<:Real}: Values of $θ$ searched using PL.

threshold::Real: threshold found using find_threshold. If pl_const was used in PL, then also add pl_const to threshold.

Returns

res: Intercepts of the threshold and $χ^2$ which corresponds to the confidence interval.

Internals

Documentation for ProfileLikelihood.jl's private functions which are not intended for use.

ProfileLikelihood.generate_incidence_data — Method

generate_incidence_data(sol::AbstractVector{<:Real})

This generates incidence data from cumulative data.

Let $\mathcal{I}:[t_0,t_1,\dots,t_n] \rightarrow \mathbb{R}$ be the number of incidences defined by $0$ if $t=t_0$ and $\mathcal{C}(t_i) - \mathcal{C}(t_{i-1})$ if $t \neq t_0$ where $\mathcal{C}$ is cumulative data.

Arguments

sol::AbstractVector{<:Real}: Vector of cumulative data.

Returns

incidence_data: Incidence data calculated from cumulative data.

ProfileLikelihood.min_point — Function

min_point(param_index::Integer, loss::Real, param_fitted::AbstractVector{<:Real})

This returns the minimum point of the profile likelihood plot where the point is (indexth element of param_fitted,loss).

ProfileLikelihood.go_right_PL — Function

go_right_PL(step_size::Real, max_steps::Integer, param_index::Integer, 
            param_fitted::AbstractVector{<:Real}, 
            data::AbstractVector{<:AbstractVector{<:Real}}, 
            sol_obs::AbstractVector{<:Integer}, 
            threshold::Real, loss::Real, 
            prob::SciMLBase.AbstractDEProblem, 
            alg_diff::SciMLBase.AbstractDEAlgorithm, 
            times::AbstractVector{<:Real}, 
            obj_arr::AbstractVector{<:Function}, 
            alg_opti, 
            lb::AbstractVector{<:Real}, ub::AbstractVector{<:Real}; 
            incidence_obs::AbstractVector{<:Integer}=Int64[], 
            solver_diff_opts::Dict=Dict(), opti_prob_opts::Dict=Dict(), 
            opti_solver_opts::Dict=Dict(), print_status::Bool=false)

This compute the points to the right of the minimum point of the profile likelihood.

Arguments

step_size::Real: Step size to take for computing points.

max_steps::Integer: Maximum steps allowed.

param_index::Integer: Index of the parameter vector that is fixed.

param_fitted::AbstractVector{<:Real}: Optimized parameters. Should be the return vector minimizer of estimate_params or estimate_params_multistart.

data::AbstractVector{<:AbstractVector{<:Real}}: Vector of data used for optimizing parameters. data must be in the same order as in sol_obs and incidence_obs.

sol_obs::AbstractVector{<:Integer}: Indices of the state variables of the DEs to be used for sampling data points.

threshold::Real: Value for which the algorithm will stop computing points if the loss of the computed points is greater than the threshold. Does not necessarily need to be the same threshold as the threshold value in find_threshold.

loss::Real: Loss of the fitted parameters according to the objective function(s). Should be the return value minimum of estimate_params or estimate_params_multistart.

prob::SciMLBase.AbstractDEProblem: DE Problem (see DifferentialEquations.jl for more information).

alg_diff::SciMLBase.AbstractDEAlgorithm: DE Solver Algorithm (see DifferentialEquations.jl for more information).

times::AbstractVector{<:Real}: Times that the data points will be sampled at.

obj_arr::AbstractVector{<:Function}: Vector of objective functions.

alg_opti: Optimization algorithm (see Optimization.jl for a list of algorithms that could be used).

lb::AbstractVector{<:Real}: Lower bound (does not need to be changed if param_index and param_eval are used).

ub::AbstractVector{<:Real}: Upper bound (does not need to be changed if param_index and param_eval are used).

Keywords

incidence_obs::AbstractVector{<:Integer}=Int64[]: Indices of the state variables of the DEs to find incidence data of. The state variables must be cumulative data.

solver_diff_opts::Dict=Dict(): Keyword arguments to be passed into the DE solver. See DifferentialEquations.jl's Common Solver Options.

opti_prob_opts::Dict=Dict(): Keyword arguments to be passed into the optimization problem. See Optimization.jl's Defining OptimizationProblems.

opti_solver_opts::Dict=Dict(): Keyword arguments to be passed into the optimization solver. See Optimization.jl's Common Solver Options.

print_status::Bool=false: Determine whether the orginal output of the optimization algorithm is printed or not.

Returns

theta_right: Values of the fixed parameter that is explored to the right of the minimum point.

sol_right: Value of likelihood computed at the fixed parameter where the other parameters are optimized.

ProfileLikelihood.go_left_PL — Function

go_left_PL(step_size::Real, max_steps::Integer, param_index::Integer, 
           param_fitted::AbstractVector{<:Real}, 
           data::AbstractVector{<:AbstractVector{<:Real}}, 
           sol_obs::AbstractVector{<:Integer}, 
           threshold::Real, loss::Real, 
           prob::SciMLBase.AbstractDEProblem, 
           alg_diff::SciMLBase.AbstractDEAlgorithm, 
           times::AbstractVector{<:Real}, 
           obj_arr::AbstractVector{<:Function}, 
           alg_opti, 
           lb::AbstractVector{<:Real}, ub::AbstractVector{<:Real}; 
           incidence_obs::AbstractVector{<:Integer}=Int64[],  
           solver_diff_opts::Dict=Dict(), opti_prob_opts::Dict=Dict(), 
           opti_solver_opts::Dict=Dict(), print_status::Bool=false)

This compute the points to the left of the minimum point of the profile likelihood.

Arguments

step_size::Real: Step size to take for computing points.

max_steps::Integer: Maximum steps allowed.

param_index::Integer: Index of the parameter vector that is fixed.

param_fitted::AbstractVector{<:Real}: Optimized parameters. Should be the return vector minimizer of estimate_params or estimate_params_multistart.

data::AbstractVector{<:AbstractVector{<:Real}}: Vector of data used for optimizing parameters. data must be in the same order as in sol_obs and incidence_obs.

sol_obs::AbstractVector{<:Integer}: Indices of the state variables of the DEs to be used for sampling data points.

threshold::Real: Value for which the algorithm will stop computing points if the loss of the computed points is greater than the threshold. Does not necessarily need to be the same threshold as the threshold value in find_threshold.

loss::Real: Loss of the fitted parameters according to the objective function(s). Should be the return value minimum of estimate_params or estimate_params_multistart.

prob::SciMLBase.AbstractDEProblem: DE Problem (see DifferentialEquations.jl for more information).

alg_diff::SciMLBase.AbstractDEAlgorithm: DE Solver Algorithm (see DifferentialEquations.jl for more information).

times::AbstractVector{<:Real}: Times that the data points will be sampled at.

obj_arr::AbstractVector{<:Function}: Vector of objective functions.

alg_opti: Optimization algorithm (see Optimization.jl for a list of algorithms that could be used).

lb::AbstractVector{<:Real}: Lower bound (does not need to be changed if param_index and param_eval are used).

ub::AbstractVector{<:Real}: Upper bound (does not need to be changed if param_index and param_eval are used).

Keywords

incidence_obs::AbstractVector{<:Integer}=Int64[]: Indices of the state variables of the DEs to find incidence data of. The state variables must be cumulative data.

solver_diff_opts::Dict=Dict(): Keyword arguments to be passed into the DE solver. See DifferentialEquations.jl's Common Solver Options.

opti_prob_opts::Dict=Dict(): Keyword arguments to be passed into the optimization problem. See Optimization.jl's Defining OptimizationProblems.

opti_solver_opts::Dict=Dict(): Keyword arguments to be passed into the optimization solver. See Optimization.jl's Common Solver Options.

print_status::Bool=false: Determine whether the orginal output of the optimization algorithm is printed or not.

Returns

theta_left: Values of the fixed parameter that is explored to the left of the minimum point.

sol_left: Value of likelihood computed at the fixed parameter where the other parameters are optimized.

ProfileLikelihood.go_right_PL_multistart — Function

go_right_PL_multistart(step_size::Real, max_steps::Integer, param_index::Integer, 
                       param_fitted::AbstractVector{<:Real}, 
                       data::AbstractVector{<:AbstractVector{<:Real}}, 
                       sol_obs::AbstractVector{<:Integer}, 
                       threshold::Real, loss::Real, 
                       prob::SciMLBase.AbstractDEProblem, 
                       alg_diff::SciMLBase.AbstractDEAlgorithm, 
                       times::AbstractVector{<:Real}, 
                       obj_arr::AbstractVector::{<:Function}, 
                       alg_opti, alg_opti_local
                       lb::AbstractVector{<:Real}, ub::AbstractVector{<:Real}; 
                       incidence_obs::AbstractVector{<:Integer}=Int64[], 
                       solver_diff_opts::Dict=Dict(), opti_prob_opts::Dict=Dict(), 
                       opti_solver_opts::Dict=Dict(), print_status::Bool=false)

This compute the points to the right of the minimum point of the profile likelihood.

This is used by find_profile_likelihood_multistart.

Arguments

step_size::Real: Step size to take for computing points.

max_steps::Integer: Maximum steps allowed.

param_index::Integer: Index of the parameter vector that is fixed.

param_fitted::AbstractVector{<:Real}: Optimized parameters. Should be the return vector minimizer of estimate_params or estimate_params_multistart.

data::AbstractVector{<:AbstractVector{<:Real}}: Vector of data used for optimizing parameters. data must be in the same order as in sol_obs and incidence_obs.

sol_obs::AbstractVector{<:Integer}: Indices of the state variables of the DEs to be used for sampling data points.

threshold::Real: Value for which the algorithm will stop computing points if the loss of the computed points is greater than the threshold. Does not necessarily need to be the same threshold as the threshold value in find_threshold.

loss::Real: Loss of the fitted parameters according to the objective function(s). Should be the return value minimum of estimate_params or estimate_params_multistart.

prob::SciMLBase.AbstractDEProblem: DE Problem (see DifferentialEquations.jl for more information).

alg_diff::SciMLBase.AbstractDEAlgorithm: DE Solver Algorithm (see DifferentialEquations.jl for more information).

times::AbstractVector{<:Real}: Times that the data points will be sampled at.

obj_arr::AbstractVector{<:Function}: Vector of objective functions.

alg_opti: Global optimization algorithm. Typically, MultistartOptimization.TikTak(n) where n is the number of starting points generated from the Sobol sequence.

alg_opti_local: Local optimization algorithm. Must be an algorithm from NLopt.jl.

lb::AbstractVector{<:Real}: Lower bound (does not need to be changed if param_index and param_eval are used).

ub::AbstractVector{<:Real}: Upper bound (does not need to be changed if param_index and param_eval are used).

Keywords

incidence_obs::AbstractVector{<:Integer}=Int64[]: Indices of the state variables of the DEs to find incidence data of. The state variables must be cumulative data.

solver_diff_opts::Dict=Dict(): Keyword arguments to be passed into the DE solver. See DifferentialEquations.jl's Common Solver Options.

opti_prob_opts::Dict=Dict(): Keyword arguments to be passed into the optimization problem. See Optimization.jl's Defining OptimizationProblems.

opti_solver_opts::Dict=Dict(): Keyword arguments to be passed into the optimization solver. See Optimization.jl's Common Solver Options.

print_status::Bool=false: Determine whether the orginal output of the optimization algorithm is printed or not.

Returns

theta_right: Values of the fixed parameter that is explored to the right of the minimum point.

sol_right: Value of likelihood computed at the fixed parameter where the other parameters are optimized.

ProfileLikelihood.go_left_PL_multistart — Function

go_left_PL_multistart(step_size::Real, max_steps::Integer, param_index::Integer, 
                      param_fitted::AbstractVector{<:Real}, 
                      data::AbstractVector{<:AbstractVector{<:Real}}, 
                      sol_obs::AbstractVector{<:Integer}, 
                      threshold::Real, loss::Real, 
                      prob::SciMLBase.AbstractDEProblem, 
                      alg_diff::SciMLBase.AbstractDEAlgorithm, 
                      times::AbstractVector{<:Real}, 
                      obj_arr::AbstractVector{<:Function}, 
                      alg_opti, alg_opti_local
                      lb::AbstractVector{<:Real}, ub::AbstractVector{<:Real}; 
                      incidence_obs::AbstractVector{<:Integer}=Int64[], 
                      solver_diff_opts::Dict=Dict(), opti_prob_opts::Dict=Dict(), 
                      opti_solver_opts::Dict=Dict(), print_status::Bool=false)

This compute the points to the left of the minimum point of the profile likelihood.

This is used by find_profile_likelihood_multistart.

Arguments

step_size::Real: Step size to take for computing points.

max_steps::Integer: Maximum steps allowed.

param_index::Integer: Index of the parameter vector that is fixed.

param_fitted::AbstractVector{<:Real}: Optimized parameters. Should be the return vector minimizer of estimate_params or estimate_params_multistart.

data::AbstractVector{<:AbstractVector{<:Real}}: Vector of data used for optimizing parameters. data must be in the same order as in sol_obs and incidence_obs.

sol_obs::AbstractVector{<:Integer}: Indices of the state variables of the DEs to be used for sampling data points.

threshold::Real: Value for which the algorithm will stop computing points if the loss of the computed points is greater than the threshold. Does not necessarily need to be the same threshold as the threshold value in find_threshold.

loss::Real: Loss of the fitted parameters according to the objective function(s). Should be the return value minimum of estimate_params or estimate_params_multistart.

prob::SciMLBase.AbstractDEProblem: DE Problem (see DifferentialEquations.jl for more information).

alg_diff::SciMLBase.AbstractDEAlgorithm: DE Solver Algorithm (see DifferentialEquations.jl for more information).

times::AbstractVector{<:Real}: Times that the data points will be sampled at.

obj_arr::AbstractVector{<:Function}: Vector of objective functions.

alg_opti: Global optimization algorithm. Typically, MultistartOptimization.TikTak(n) where n is the number of starting points generated from the Sobol sequence.

alg_opti_local: Local optimization algorithm. Must be an algorithm from NLopt.jl.

lb::AbstractVector{<:Real}: Lower bound (does not need to be changed if param_index and param_eval are used).

ub::AbstractVector{<:Real}: Upper bound (does not need to be changed if param_index and param_eval are used).

Keywords

incidence_obs::AbstractVector{<:Integer}=Int64[]: Indices of the state variables of the DEs to find incidence data of. The state variables must be cumulative data.

solver_diff_opts::Dict=Dict(): Keyword arguments to be passed into the DE solver. See DifferentialEquations.jl's Common Solver Options.

opti_prob_opts::Dict=Dict(): Keyword arguments to be passed into the optimization problem. See Optimization.jl's Defining OptimizationProblems.

opti_solver_opts::Dict=Dict(): Keyword arguments to be passed into the optimization solver. See Optimization.jl's Common Solver Options.

print_status::Bool=false: Determine whether the orginal output of the optimization algorithm is printed or not.

Returns

theta_right: Values of the fixed parameter that is explored to the left of the minimum point.

sol_right: Value of likelihood computed at the fixed parameter where the other parameters are optimized.