Docstrings

Here we give some of the main docstrings.

LikelihoodProblem

ProfileLikelihood.AbstractLikelihoodProblem — Type

abstract type AbstractLikelihoodProblem{N, L}

Abstract type of a likelihood problem, where N is the number of parameters and L is the type of the likelihood function.

ProfileLikelihood.LikelihoodProblem — Type

LikelihoodProblem{N,P,D,L,Θ,S} <: AbstractLikelihoodProblem

Struct representing a likelihood problem.

Fields

problem::P

The associated OptimizationProblem.

data::D

The argument p used in the log-likelihood function.

log_likelihood_function::L

The log-likelihood function, taking the form ℓ(θ, p).

θ₀::Θ

Initial estimates for the MLE θ.

syms::S

Variable names for the parameters.

The extra parameter N is the number of parameters.

Constructors

Standard

LikelihoodProblem(loglik::Function, θ₀;
    syms=eachindex(θ₀), data=SciMLBase.NullParameters(),
    f_kwargs=nothing, prob_kwargs=nothing)

Constructor for the LikelihoodProblem.

Arguments

loglik::Function: The log-likelihood function, taking the form ℓ(θ, p).

θ₀: The estimates estimates for the MLEs.

Keyword Arguments

syms=eachindex(θ₀): Names for each parameter.

data=SciMLBase.NullParameters(): The parameter p in the log-likelihood function.

f_kwargs=nothing: Keyword arguments, passed as a NamedTuple, for the OptimizationFunction.

prob_kwargs=nothing: Keyword arguments, passed as a NamedTuple, for the OptimizationProblem.

Outputs

Returns the LikelihoodProblem problem object.

With arguments for a differential equation problem

LikelihoodProblem(loglik::Function, θ₀,
    ode_function, u₀, tspan;
    syms=eachindex(θ₀), data=SciMLBase.NullParameters(),
    ode_parameters=SciMLBase.NullParameters(), ode_alg,
    ode_kwargs=nothing, f_kwargs=nothing, prob_kwargs=nothing)

Constructor for the LikelihoodProblem for a differential equation problem.

Arguments

loglik::Function: The log-likelihood function, taking the form ℓ(θ, p, integrator).

θ₀: The estimates estimates for the MLEs.

ode_function: The function f(du, u, p, t) or f(u, p, t) for the differential equation.

u₀: The initial condition for the differential equation.

tspan: The time-span to solve the differential equaton over.

Keyword Arguments

syms=eachindex(θ₀): Names for each parameter.

data=SciMLBase.NullParameters(): The parameter p in the log-likelihood function.

ode_parameters=SciMLBase.NullParameters(): The parameter p in ode_function.

ode_alg: The algorithm used for solving the differential equatios.

ode_kwargs=nothing: Extra keyword arguments, passed as a NamedTuple, to pass into the integrator; see construct_integrator.

f_kwargs=nothing: Keyword arguments, passed as a NamedTuple, for the OptimizationFunction.

prob_kwargs=nothing: Keyword arguments, passed as a NamedTuple, for the OptimizationProblem.

Outputs

Returns the LikelihoodProblem problem object.

With an integrator

LikelihoodProblem(loglik::Function, θ₀, integrator;
    syms=eachindex(θ₀), data=SciMLBase.NullParameters(),
    f_kwargs=nothing, prob_kwargs=nothing)

Constructor for the LikelihoodProblem for a differential equation problem with associated integrator.

Arguments

loglik::Function: The log-likelihood function, taking the form ℓ(θ, p, integrator).

θ₀: The estimates estimates for the MLEs.

integrator: The integrator for the differential equation problem. See also construct_integrator.

Keyword Arguments

syms=eachindex(θ₀): Names for each parameter.

data=SciMLBase.NullParameters(): The parameter p in the log-likelihood function.

f_kwargs=nothing: Keyword arguments, passed as a NamedTuple, for the OptimizationFunction.

prob_kwargs=nothing: Keyword arguments, passed as a NamedTuple, for the OptimizationProblem.

Outputs

Returns the LikelihoodProblem problem object.

LikelihoodSolution

ProfileLikelihood.AbstractLikelihoodSolution — Type

abstract type AbstractLikelihoodSolution{N, P}

Type representing the solution to a likelihood problem, where N is the number of parameters and P is the type of the likelihood problem.

ProfileLikelihood.LikelihoodSolution — Type

struct LikelihoodSolution{Θ,P,M,R,A} <: AbstractLikelihoodSolution

Struct for a solution to a LikelihoodProblem.

Fields

mle::Θ: The MLEs.

problem::P: The LikelihoodProblem.

optimiser::A: The algorithm used for solving the optimisation problem.

maximum::M: The maximum likelihood.

retcode::R: The SciMLBase.ReturnCode.

Constructors

LikelihoodSolution(sol::SciMLBase.OptimizationSolution, prob::AbstractLikelihoodProblem; alg=sol.alg)

Constructs the likelihood solution from a solution to an OptimizationProblem with a given LikelihoodProblem.

ProfileLikelihood.mle — Function

mle(prob::LikelihoodProblem, alg, args...; kwargs...)
mle(prob::LikelihoodProblem, alg::Tuple, args...; kwargs...)

Given the likelihood problem prob and an optimiser alg, finds the MLEs and returns a LikelihoodSolution object. Extra arguments and keyword arguments for solve can be passed through args... and kwargs....

If alg is a Tuple, then the problem is re-optimised after each algorithm with the next element in alg, starting from alg[1], with initial estimate coming from the solution with the previous algorithm (starting with get_initial_estimate(prob)).

ProfileLikelihoodSolution

ProfileLikelihood.ProfileLikelihoodSolution — Type

ProfileLikelihoodSolution{I,V,LP,LS,Spl,CT,CF,OM}

Struct for the normalised profile log-likelihood. See profile for a constructor.

Fields

θ::Dict{I, V}: This is a dictionary such that θ[i] gives the parameter values used for the normalised profile log-likelihood of the ith variable.

profile::Dict{I, V}: This is a dictionary such that profile[i] gives the values of the normalised profile log-likelihood function at the corresponding values in θ[i].

prob::LP: The original LikelihoodProblem.

mle::LS: The solution to the full problem.

spline::Dict{I, Spl}: This is a dictionary such that spline[i] is a spline through the data (θ[i], profile[i]). This spline can be evaluated at a point ψ for the ith variable by calling an instance of the struct with arguments (ψ, i). See also spline_profile.

confidence_intervals::Dict{I,ConfidenceInterval{CT,CF}}: This is a dictonary such that confidence_intervals[i] is a confidence interval for the ith parameter.

other_mles::OM: This is a dictionary such that other_mles[i] gives the vector for the MLEs of the other parameters not being profiled, for each datum.

Spline evaluation

This struct is callable. We define the method

(prof::ProfileLikelihoodSolution)(θ, i)

that evaluates the spline through the ith profile at the point θ.

ProfileLikelihood.ConfidenceInterval — Type

struct ConfidenceInterval{T, F}

Struct representing a confidence interval.

Fields

lower::T: The lower bound of the confidence interval.

upper::T: The upper bound of the confidence interval.

level::F: The level of the confidence interval.

ProfileLikelihood.profile — Function

profile(prob::LikelihoodProblem, sol::LikelihoodSolution, n=1:number_of_parameters(prob);
    alg=get_optimiser(sol),
    conf_level::F=0.95,
    confidence_interval_method=:spline,
    threshold=get_chisq_threshold(conf_level),
    resolution=200,
    param_ranges=construct_profile_ranges(sol, get_lower_bounds(prob), get_upper_bounds(prob), resolution),
    min_steps=10,
    normalise::Bool=true,
    spline_alg=FritschCarlsonMonotonicInterpolation,
    extrap=Line,
    parallel=false,
    next_initial_estimate_method = :prev,
    kwargs...)

Computes profile likelihoods for the parameters from a likelihood problem prob with MLEs sol.

See also replace_profile! which allows you to re-profile a parameter in case you are not satisfied with the results. For plotting, see the plot_profiles function (requires that you have loaded CairoMakie.jl and LaTeXStrings.jl to access the function).

Arguments

prob::LikelihoodProblem: The LikelihoodProblem.

sol::LikelihoodSolution: The LikelihoodSolution. See also mle.

n=1:number_of_parameters(prob): The parameter indices to compute the profile likelihoods for.

Keyword Arguments

alg=get_optimiser(sol): The optimiser to use for solving each optimisation problem.

conf_level::F=0.95: The level to use for the ConfidenceIntervals.

confidence_interval_method=:spline: The method to use for computing the confidence intervals. See also get_confidence_intervals!. The default :spline uses rootfinding on the spline through the data, defining a continuous function, while the alternative :extrema simply takes the extrema of the values that exceed the threshold.

threshold=get_chisq_threshold(conf_level): The threshold to use for defining the confidence intervals.

resolution=200: The number of points to use for evaluating the profile likelihood in each direction starting from the MLE (giving a total of 400 points).

param_ranges=construct_profile_ranges(sol, get_lower_bounds(prob), get_upper_bounds(prob), resolution): The ranges to use for each parameter.

min_steps=10: The minimum number of steps to allow for the profile in each direction. If fewer than this number of steps are used before reaching threshold, then the algorithm restarts and computes the profile likelihood a number min_steps of points in that direction.

normalise::Bool=true: Whether to optimise the normalised profile log-likelihood or not.

spline_alg=FritschCarlsonMonotonicInterpolation: The interpolation algorithm to use for computing a spline from the profile data. See Interpolations.jl.

extrap=Line: The extrapolation algorithm to use for computing a spline from the profile data. See Interpolations.jl.

parallel=false: Whether to use multithreading. If true, will use multithreading so that multiple parameters are profiled at once, and the steps to the left and right are done at the same time.

next_initial_estimate_method = :prev: Method for selecting the next initial estimate when stepping forward when profiling. :prev simply uses the previous solution, but you can also use :interp to use linear interpolation. See also set_next_initial_estimate!.

kwargs...: Extra keyword arguments to pass into solve for solving the OptimizationProblem. See also the docs from Optimization.jl.

ProfileLikelihood.replace_profile! — Function

replace_profile!(prof::ProfileLikelihoodSolution, n);
    alg=get_optimiser(prof.likelihood_solution),
    conf_level::F=0.95,
    confidence_interval_method=:spline,
    threshold=get_chisq_threshold(conf_level),
    resolution=200,
    param_ranges=construct_profile_ranges(prof.likelihood_solution, get_lower_bounds(prof.likelihood_problem), get_upper_bounds(prof.likelihood_problem), resolution),
    min_steps=10,
    normalise::Bool=true,
    spline_alg=FritschCarlsonMonotonicInterpolation,
    extrap=Line,
    parallel=false,
    next_initial_estimate_method=:prev,
    kwargs...) where {F}

Given an existing prof::ProfileLikelihoodSolution, replaces the profile results for the parameters in n. The keyword arguments are the same as for profile.

ProfileLikelihood.set_next_initial_estimate! — Function

set_next_initial_estimate!(sub_cache, param_vals, other_mles, prob, θₙ; next_initial_estimate_method=:prev)

Method for selecting the next initial estimate for the optimisers. sub_cache is the cache vector for placing the initial estimate into, param_vals is the current list of parameter values for the interest parameter, and other_mles is the corresponding list of previous optimisers. prob is the OptimizationProblem. The value θₙ is the next value of the interest parameter.

The available methods are:

next_initial_estimate_method = :prev: If this is selected, simply use other_mles[end], i.e. the previous optimiser.

next_initial_estimate_method = :interp: If this is selected, the next optimiser is determined via linear interpolation using the data (param_vals[end-1], other_mles[end-1]), (param_vals[end], other_mles[end]). If the new approximation is outside of the parameter bounds, falls back to next_initial_estimate_method = :prev.

Prediction intervals

ProfileLikelihood.get_prediction_intervals — Function

get_prediction_intervals(q, prof::ProfileLikelihoodSolution, data;
    q_type=get_q_type(q, prof, data), resolution=250)

Given a prediction of the form q(θ, data), where θ has the same size as the θ used in the profile likelihood solution prof, and data is the argument data, computes the prediction intervals for q (possibly at each point if it outputs a vector) using the confidence intervals from prof. You can set the output of the prediction function using q_type, and the grid resolution when evaluating the prediction function for each parameter via resolution.

Two results are produced:

parameterwise_cis: This is a Dict mapping parameter indices to a a vector of confidence intervals from each output of q for the corresponding parameter.

union_cis: This gives the union of the intervals from parameterwise_cis (just taking the extrema over each interal) at each output of q.

We also return all_curves, a Dict mapping parameter indices to the vector of q values for each parameter, and these parameter ranges are given in param_ranges. So, the final output looks like:

(parameterwise_cis, union_cis, all_curves, param_ranges).

ProfileLikelihood.eval_prediction_function — Function

eval_prediction_function(q, prof::ProfileLikelihoodSolution, data;
    resolution=250,
    param_ranges=Dict(profiled_parameters(prof) .=> [LinRange(get_confidence_intervals(prof[i])..., resolution) for i in profiled_parameters(prof)]),
    q_type=get_q_type(q, prof, data))

Given a prediction of the form q(θ, data), where θ has the same size as the θ used in the profile likelihood solution prof, and data is the argument data, and for each parameter index i: Evaluates q((ψ, ωˢ(ψ)), data), where ψ ranges over param_ranges[i] and ωˢ(ψ) are the parameter values that lead to the value of the profile likelihood function where θ[i] = ψ.

You can set the type of the output from the prediction function q using q_type.

The output is a Dict mapping the profiled parameter indices (from profiled_parameters) to the outputs from q at each corresponding parameter in param_ranges.

Plotting

ProfileLikelihood.plot_profiles — Function

plot_profiles(prof::ProfileLikelihoodSolution, vars = profiled_parameters(prof); 
    ncol=nothing, 
    nrow=nothing,
    true_vals=Dict(vars .=> nothing), 
    spline=true, 
    show_mles=true, 
    shade_ci=true, 
    fig_kwargs=nothing, 
    axis_kwargs=nothing,
    latex_names = Dict(vars .=> [LaTeXStrings.L"	heta_{i}" for i in SciMLBase.sym_to_index.(vars, Ref(prof))]))

Plot results from a profile likelihood solution prof.

Arguments

prof::ProfileLikelihoodSolution: The profile likelihood solution from profile.

vars = profiled_parameters(prof): The parameters to plot.

Keyword Arguments

ncol=nothing: The number of columns to use. If nothing, chosen automatically via choose_grid_layout.

nrow=nothing: The number of rows to use. If nothing, chosen automatically via choose_grid_layout

true_vals=Dict(vars .=> nothing): A dictionary mapping parameter indices to their true values, if they exist. If nothing, nothing is plotted, otherwise a black line is plotted at the true value for the profile.

spline=true: Whether the curve plotted should come from a spline through the results, or if the data itself should be plotted.

show_mles=true: Whether to put a red line at the MLEs.

shade_ci=true: Whether to shade the area under the profile between the confidence interval.

fig_kwargs=nothing: Extra keyword arguments for Figure (see the Makie docs).

axis_kwargs=nothing: Extra keyword arguments for Axis (see the Makie docs).

latex_names = Dict(vars .=> [LaTeXStrings.L" heta_{i}" for i in SciMLBase.sym_to_index.(vars, Ref(prof))])): LaTeX names to use for the parameters. Defaults to θᵢ, where i is the index of the parameter.

Output

The Figure() is returned.

GridSearch

Grid definitions

ProfileLikelihood.AbstractGrid — Type

abstract type AbstractGrid{N,B,T}

Type representing a grid, where N is the number of parameters, B is the type for the bounds, and T is the number type.

ProfileLikelihood.RegularGrid — Type

struct RegularGrid{N,B,R,S,T} <: AbstractGrid{N,B,T}

Struct for a grid in which each parameter is regularly spaced.

Fields

lower_bounds::B: Lower bounds for each parameter.

upper_bounds::B: Upper bounds for each parameter.

resolution::R: Number of grid points for each parameter. If R <: Number, then the same number of grid points is used for each parameter.

step_sizes::S: Grid spacing for each parameter.

Constructor

You can construct a RegularGrid using RegularGrid(lower_bounds, upper_bounds, resolution).

ProfileLikelihood.IrregularGrid — Type

struct IrregularGrid{N,B,R,S,T} <: AbstractGrid{N,B,T}

Struct for an irregular grid of parameters.

Fields

lower_bounds::B: Lower bounds for each parameter.

upper_bounds::B: Upper bounds for each parameter.

grid::G: The set of parameter values, e.g. a matrix where each column is the parameter vector.

Constructor

You can construct a IrregularGrid using IrregularGrid(lower_bounds, upper_bounds, grid).

Performing a grid search

ProfileLikelihood.GridSearch — Type

struct GridSearch{F,G}

Struct for a GridSearch.

Fields

f::F: The function to optimise, of the form f(x, p).

p::P: The arguments p in the function f.

grid::G: The grid, where G<:AbstractGrid. See also grid_search.

ProfileLikelihood.grid_search — Function

grid_search(prob; save_vals=Val(false), minimise:=Val(false), parallel=Val(false))

Performs a grid search for the given grid search problem prob.

Arguments

prob::GridSearch{F, G}: The grid search problem.

Keyword Arguments

save_vals:=Val(false): Whether to return a array with the function values.

minimise:=Val(false): Whether to minimise or to maximise the function.

parallel:=Val(false): Whether to run the grid search with multithreading.

Outputs

f_opt: The optimal objective value.

x_argopt: The parameter that gave f_opt.

f_res: If save_vals==Val(true), then this is the array of function values.

grid_search(f, grid::AbstractGrid; save_vals=Val(false), minimise=Val(false), parallel=Val(false))

For a given grid and function f, performs a grid search.

Arguments

f: The function to optimise.

grid::AbstractGrid: The grid to use for optimising.

Keyword Arguments

save_vals=Val(false): Whether to return a array with the function values.

minimise=Val(false): Whether to minimise or to maximise the function.

parallel=Val(false): Whether to run the grid search with multithreading.

grid_search(prob::LikelihoodProblem, grid::AbstractGrid, parallel=Val(false); save_vals=Val(false))

Given a grid and a likelihood problem prob, maximises it over the grid using a grid search. If save_vals==Val(true), then the likelihood function values at each gridpoint are returned. Set parallel=Val(true) if you want multithreading.